\name{confint.ivmod}

\alias{confint.ivmod}

\title{
Confidence interval
}

\description{
This is a \code{confint} method for class \code{"ivmod"}.  
}

\usage{
\method{confint}{ivmod}(object, parm, level=0.95, ...)
}

\arguments{
  \item{object}{
an object of class \code{"ivmod"}.  
} 
  \item{parm}{
not used.
}
  \item{level}{
the coverage probability of the confidence intervals. 
}
  \item{\dots}{
not used.
}

}

\author{
Arvid Sjolander. 
}

