\name{plot.estfun}

\alias{plot.estfun}

\title{
Plots sums of estimating functions. 
}

\description{
This is a \code{plot} method for class \code{"estfun"}. 
}

\usage{
\method{plot}{estfun}(x, ...)
}

\arguments{
  \item{x}{
an object of class \code{"estfun"}, fitted with \code{method="g"}.
}  
 \item{\dots}{
additional arguments to \code{plot}.
}
}

\author{
Arvid Sjolander.
}

\examples{

##See documentation for estfun.

}
