\name{print.ivmod}

\alias{print.ivmod}

\title{
Prints output of instrumental variable estimation
}

\description{
This is a \code{print} method for class \code{"ivmod"}.
}

\usage{
\method{print}{ivmod}(x, digits=max(3L, getOption("digits")-3L), ...)
}

\arguments{
  \item{x}{
an object of class \code{"ivmod"}.
}
  \item{digits}{
the number of significant digits to use.
}
  \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}


\examples{

##See documentation for ivglm, ivcoxph and ivah.

}
