\name{print.summary.ivbounds}

\alias{print.summary.ivbounds}

\title{
Prints summary of instrumental variable bounds
}

\description{
This is a \code{print} method for class \code{"summary.ivbounds"}.
}

\usage{
\method{print}{summary.ivbounds}(x, digits=max(3L, getOption("digits")-3L), 
   ...)
}

\arguments{
  \item{x}{
an object of class \code{"summary.ivbounds"}.
}
  \item{digits}{
the number of significant digits to use.
}
  \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}


\examples{

##See documentation for ivbounds.

}
