\name{summary.ivmod}

\alias{summary.ivmod}

\title{
Summarizes instrumental variable estimation
}

\description{
This is a \code{summary} method for class \code{"ivmod"}.
}

\usage{
\method{summary}{ivmod}(object, digits=max(3L, getOption("digits")-3L), ...)
}

\arguments{
  \item{object}{
an object of class \code{"ivmod"}.
} 
  \item{digits}{
the number of significant digits to use.
}  
 \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}

\examples{

##See documentation for ivglm, ivcoxph and ivah.

}
