\name{summary.tsls}

\alias{summary.tsls}

\title{
Summarizes two-stage least squares estimation
}

\description{
This is a \code{summary} method for class \code{"tsls"}.
}

\usage{
\method{summary}{tsls}(object, digits=max(3L, getOption("digits")-3L), ...)
}

\arguments{
  \item{object}{
an object of class \code{"tsls"}.
} 
  \item{digits}{
the number of significant digits to use.
}  
 \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}

\seealso{
\code{\link{tsls}}
}

\examples{

##See documentation for tsls

}
